# api-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/user": {
      "get": {
        "operationId": "AuthController_user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The authenticated user.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User was successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Cannot log in the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot log in the user. Invalid data was provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": "User refresh and access tokens were successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot update user's refresh and access tokens. Invalid data was provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "delete": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The user was successfully logged out.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseEntity"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserPublicEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "403": {
            "description": "The user is forbidden to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "409": {
            "description": "Cannot log out the user. Invalid data was provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error was occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-Commerce Order Management API",
    "description": "RESTful API for managing users and orders in an e-commerce system",
    "version": "1.0.0",
    "contact": {
      "name": "Contact API Support",
      "url": "https://t.me/kalyonekenobe",
      "email": "oleksandr.ihumnov@ukma.edu.ua"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.ukmashop.com/api/v1",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponseEntity": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": "Invalid email"
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "UserPublicEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user (UUID v4)",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "description": "User email address used for login and communication",
            "format": "email",
            "example": "user@ukmashop.com"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "phone": {
            "type": "object",
            "description": "User phone number in international format (10–15 digits, optional + prefix)",
            "example": "+1234567890"
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user that determines access permissions",
            "enum": [
              "User",
              "Admin",
              "Manager"
            ],
            "example": "User"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates whether the user account is currently active",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the user account was created (ISO 8601 format)",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date and time when the user account was last updated (ISO 8601 format)",
            "example": "2024-01-20T14:45:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phone",
          "role",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApiErrorResponseEntity": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-02-15T21:24:28.191Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/endpoint"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occured on server"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "example": null,
            "additionalProperties": {}
          },
          "errors": {
            "type": "object",
            "example": {
              "server": [
                "An error occured on server"
              ]
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "timestamp",
          "path",
          "success",
          "message",
          "data",
          "errors"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "object",
            "description": "The email of the user",
            "example": "oleksandr.igumnov@gmail.com",
            "format": "email"
          },
          "password": {
            "type": "object",
            "description": "User's password",
            "example": "Password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}